services/pipeline-backend/pyproject.toml
[project]
name = "build-commit-pipeline-backend"
version = "0.1.0"
description = "FastAPI + Celery backend for the TravisTorrent SonarQube pipeline"
requires-python = ">=3.11"
dependencies = [
    "buildguard-common",
    "pydantic==2.10.1",
    "pydantic-settings==2.6.1",
    "motor==3.6.0",
    "fastapi==0.115.5",
    "uvicorn==0.32.1",
    "python-multipart==0.0.9",
    "celery==5.4.0",
    "redis==5.2.0",
    "requests==2.32.3",
    "httpx==0.27.2",
    "pandas==2.2.3",
    "pyyaml==6.0.2",
    "boto3==1.35.71",
    "opentelemetry-distro",
    "opentelemetry-exporter-zipkin",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-celery",
    "opentelemetry-instrumentation-pymongo",
    "opentelemetry-instrumentation-requests",
    "opentelemetry-instrumentation-logging",
    "pybreaker==1.3.0",
    "python-json-logger==2.0.7",
    "GitPython>=3.1,<3.2",
    "PyGithub>=2.1,<2.3",
    "python-jose[cryptography]>=3.3,<3.4",
    "python-dateutil>=2.8,<3.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv.sources]
buildguard-common = { workspace = true }


[tool.setuptools.packages.find]
include = ["app*", "pipeline*"]
